% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\docType{data}
\name{AdapterMemory}
\alias{AdapterMemory}
\title{AdapterMemory}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AdapterMemory
}
\description{
AdapterMemory
}
\keyword{datasets}
