% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapters.R
\docType{data}
\name{FeetureAdapter}
\alias{FeetureAdapter}
\title{FeetureAdapter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FeetureAdapter
}
\description{
FeetureAdapter
}
\keyword{datasets}
