% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feetures-package.R
\docType{package}
\name{feetures-package}
\alias{feetures-package}
\alias{feetures}
\title{feetures}
\description{
feetures
}
\author{
Scott Chamberlain \email{sckott@protonmail.com}
}
\keyword{package}
